Rescue after_error callback errors
If a job fails currently, it attempts an `after_error/2` callback.

If that callback exists it is run. If it doesn't the dispatcher moves
on.

This commit handles the case where the callback is there, but it raises
an error. Previously if this happened, the dispatcher would die and the
job would be put back on the queue and picked up again when a new
dispatcher was created.