Add a new `priority` column
Defining a default priority
---------------------------

It makes sense to provide a default priority where possible. This way
the implementation is backwards compatible for any users who are
implementing their own jobs with `Rihanna.Job`.

The lowest priority is the default, allowing developers to increase
the priority of their jobs as/when they need to.

The default priority value
--------------------------

A default of 19 has been chosen to conform with the Unix process
priority scale[1]

[1] https://en.wikipedia.org/wiki/Nice_(Unix)